Move timeline to repeater screen
